agent/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
openai==1.59.5
pydantic==2.10.4
python-dotenv==1.0.1
httpx==0.28.1
